# Simplify conditional.py