##### SUMMARY
Quick perf test:
```python
In [1]: import timeit
In [2]: from ansible.playbook.conditional import Conditional
In [3]: from ansible.template import Templar
In [4]: from ansible.parsing.dataloader import DataLoader
In [5]: t = Templar(None)
In [6]: c = Conditional(loader=DataLoader())
In [7]: c.when = ["x == y"]*100
In [8]: %timeit c.evaluate_conditional(templar=t, all_vars={"x": 1, "y": 1})
```
Before:
```
89.2 ms ± 1.5 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After:
```
58.9 ms ± 820 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

ci_complete

##### ISSUE TYPE
<!--- Pick one below and delete the rest -->
- Bugfix Pull Request

##### COMPONENT NAME
<!--- Write the short name of the module, plugin, task or feature below -->
`lib/ansible/playbook/conditional.py`